added lang file functionality

+++ src/main/java/theqwerdev/custommusicdiscs/item/MusicDiscAdder.java
@@ -6,16 +6,14 @@ import theqwerdev.custommusicdiscs.CustomMusicDiscsMod;
 import turniplabs.halplibe.helper.ItemHelper;
 import turniplabs.halplibe.helper.SoundHelper;
 
-import java.io.File;
-import java.io.FileWriter;
-import java.io.IOException;
+import java.io.*;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
 import java.util.ArrayList;
+import java.util.Objects;
 
 public class MusicDiscAdder {
-
 	static int startingID = CustomMusicDiscsConfig.itemID;
 	static Path musicPath = Paths.get(CustomMusicDiscsConfig.musicPath);
 	static String[] exts = {".ogg", ".wav", ".mus"};
@@ -39,41 +37,37 @@ public class MusicDiscAdder {
 			});
 
 			if(musicFiles != null && musicFiles.length > 0) {
-				//CustomMusicDiscsMod.LOGGER.warn(getClass().getResource("/lang/" + CustomMusicDiscsMod.MOD_ID + "/en_US.lang").getPath().replace("//", "/").trim());
-				Path langPath = Paths.get("/lang/" + CustomMusicDiscsMod.MOD_ID + "/en_US.lang");
-				try {
-					Files.delete(langPath);
-					Files.createFile(langPath);
-				} catch (IOException e) {
-					CustomMusicDiscsMod.LOGGER.warn(e.toString());
-				}
+				if(CustomMusicDiscsMod.isClient) {
+					File langFile = new File(Objects.requireNonNull(getClass().getResource("/lang/" + CustomMusicDiscsMod.MOD_ID + "/en_US.lang")).getPath()); //yeah, this sucks
 
-				for (File musicFile : musicFiles) {
-					String name = musicFile.getName();
-					int extPos = name.lastIndexOf('.');
-					name = name.substring(0, extPos);
-
-					discs.add(ItemHelper.createItem(CustomMusicDiscsMod.MOD_ID,
-						new ItemCustomRecord("record.custom" + (discs.size() + 1), startingID + discs.size(), name), "custom" + (discs.size() + 1), "disc_custom.png"));
-
+					langFile.delete();
 					try {
-						Path tempPath = Files.copy(musicFile.toPath(), Paths.get(SoundHelper.streamingDirectory.getPath(), musicFile.getName()));
-						tempPath.toFile().deleteOnExit();
-
-						CustomMusicDiscsMod.LOGGER.info("Imported '" + musicFile.getName() + '\'');
+						langFile.createNewFile();
 					} catch (IOException e) {
 						CustomMusicDiscsMod.LOGGER.warn(e.toString());
 					}
 
-					File langFile = langPath.toFile();
-					try {
-						FileWriter fw = new FileWriter(langFile, true);
-						fw.write("item." + CustomMusicDiscsMod.MOD_ID + ".custom" + discs.size() + ".name=Custom Music Disc\n" +
-							"item." + CustomMusicDiscsMod.MOD_ID + ".custom" + discs.size() + ".desc=" + name + "\n\n");
-						//item.custommusicdiscs.church1.desc=§4Cantarile §4I §4si §4III, §4ectenia §4intreita
-						fw.close();
-					} catch (IOException e) {
-						CustomMusicDiscsMod.LOGGER.warn(e.toString());
+					for (File musicFile : musicFiles) {
+						String name = musicFile.getName();
+						int extPos = name.lastIndexOf('.');
+						name = name.substring(0, extPos);
+
+						discs.add(ItemHelper.createItem(CustomMusicDiscsMod.MOD_ID,
+							new ItemCustomRecord("record.custom" + (discs.size() + 1), startingID + discs.size(), name), "disc_gold.png"));
+
+						try {
+							Path tempPath = Files.copy(musicFile.toPath(), Paths.get(SoundHelper.streamingDirectory.getPath(), musicFile.getName()));
+							tempPath.toFile().deleteOnExit();
+
+							CustomMusicDiscsMod.LOGGER.info("Imported '" + musicFile.getName() + '\'');
+
+							FileWriter fw = new FileWriter(langFile, true);
+							fw.write("item." + CustomMusicDiscsMod.MOD_ID + ".record.custom" + discs.size() + ".name=Custom Music Disc\n" +
+								"item." + CustomMusicDiscsMod.MOD_ID + ".record.custom" + discs.size() + ".desc=" + name + "\n\n");
+							fw.close();
+						} catch (IOException e) {
+							CustomMusicDiscsMod.LOGGER.warn(e.toString());
+						}
 					}
 				}
 			}
@@ -95,11 +89,5 @@ public class MusicDiscAdder {
 				CustomMusicDiscsMod.LOGGER.warn(e.toString());
 			}
 		}
-
-		//discs.add(ItemHelper.createItem(CustomMusicDiscsMod.MOD_ID,
-			//new ItemCustomRecord("record.church1", startingID + discs.size(), "Cantarile I si III, ectenia intreita"), "church1", "disc_gold.png"));
-		//discs.add(ItemHelper.createItem(CustomMusicDiscsMod.MOD_ID,
-			//new ItemCustomRecord("record.cawdgagdadg", startingID + discs.size(), "Cantarile I si III, ectenia intreita"), "agdagdagd", "disc_gold.png"));
 	}
-
 }
